Refactor initial CLI test
